backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
torch>=2.6.0
torchvision>=0.21.0
pillow==10.2.0